requirements.txt
aiogram>=3.7.0,<4.0.0
yt-dlp>=2024.8.6
python-dotenv>=1.0.1
